+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### PaymentSheet
 * [Fixed] Fixed an issue where confirming Cash App Pay, Satispay, or PayPal could fail when setup_future_usage = `on_session`.
+* [Changed] Custom Payment Methods is now generally available, learn more by visiting the [docs](https://docs.stripe.com/payments/mobile/custom-payment-methods).
+
 
 ## 25.3.1 2025-12-19